T4.x support
Add T4 and hopefully T4.1 support to the library.

The code was built to be semi table driven starting
from the T4 PWM source table, which was extended to
include needed additional information such as INPUT mux source,
ISRs, ...

The code part was based on the FreqMeasure code that was earlier
updated to  work on T4 for only pin 22.  The code was converted
to use the underlying IMXRT FlexPWM structure and the like.
The code tries to support of the PWM pins which are FlexPWM.

The code has some complications that some of the FlexPWM objects
sub-objects have multiple channels on the (A, B, X) so the code
may chain a few of these together and run off of the same ISR

Added begin options to keywords.txt plus readLevel method

Added a few text sketches to help test, including one where I
tried to show the results from T3.6 and T4 for all of the different
capture types.